+++ constellix/resource_constellix_cname_record.go
@@ -6,6 +6,7 @@ import (
 	"io/ioutil"
 	"log"
 	"strconv"
+	"strings"
 
 	"github.com/Constellix/constellix-go-client/client"
 	"github.com/Constellix/constellix-go-client/models"
@@ -18,6 +19,10 @@ func resourceConstellixCNameRecord() *schema.Resource {
 		Update: resourceConstellixCNameRecordUpdate,
 		Read:   resourceConstellixCNameRecordRead,
 		Delete: resourceConstellixCNameRecordDelete,
+
+		Importer: &schema.ResourceImporter{
+			State: resourceConstellixCNameRecordImport,
+		},
 
 		Schema: map[string]*schema.Schema{
 			"domain_id": &schema.Schema{
@@ -155,6 +160,84 @@ func resourceConstellixCNameRecord() *schema.Resource {
 			},
 		},
 	}
+}
+
+func resourceConstellixCNameRecordImport(d *schema.ResourceData, m interface{}) ([]*schema.ResourceData, error) {
+	log.Printf("[DEBUG] %s: Beginning Import", d.Id())
+	constellixClient := m.(*client.Client)
+	params := strings.Split(d.Id(), ":")
+	resp, err := constellixClient.GetbyId("v1/" + params[0] + "/" + params[1] + "/records/cname/" + params[2])
+	if err != nil {
+		if resp.StatusCode == 404 {
+			d.SetId("")
+			return nil, err
+		}
+		return nil, err
+	}
+	bodyBytes, err := ioutil.ReadAll(resp.Body)
+	if err != nil {
+		return nil, err
+	}
+	bodyString := string(bodyBytes)
+	var data map[string]interface{}
+	err = json.Unmarshal([]byte(bodyString), &data)
+	if err != nil {
+		return nil, err
+	}
+
+	geoloc1 := data["geolocation"]
+	log.Println("GEOLOC VALUE INSIDE READ :", geoloc1)
+	geoset := make(map[string]interface{})
+	if geoloc1 != nil {
+		geoloc := geoloc1.(map[string]interface{})
+		geoset["geo_ip_user_region"], _ = strconv.Atoi(fmt.Sprintf("%v", geoloc["geoipFilter"]))
+		geoset["drop"] = fmt.Sprintf("%v", geoloc["drop"])
+		geoset["geo_ip_proximity"], _ = strconv.Atoi(fmt.Sprintf("%v", geoloc["geoipProximity"]))
+	} else {
+		geoset = nil
+	}
+
+	rcdf := data["recordFailover"]
+	rcdfset := make(map[string]interface{})
+	rcdflist := make([]interface{}, 0, 1)
+	if rcdf != nil {
+		rcdf1 := rcdf.(map[string]interface{})
+		rcdfset["record_failover_failover_type"] = fmt.Sprintf("%v", rcdf1["failoverType"])
+		rcdfset["record_failover_disable_flag"] = fmt.Sprintf("%v", rcdf1["disabled"])
+
+		rcdfvalues := rcdf1["values"].([]interface{})
+		log.Println("LIST VALUES: ", rcdfvalues)
+
+		for _, valrcdf := range rcdfvalues {
+			map1 := make(map[string]interface{})
+			val1 := valrcdf.(map[string]interface{})
+			map1["value"] = fmt.Sprintf("%v", val1["value"])
+			map1["sort_order"] = fmt.Sprintf("%v", val1["sortOrder"])
+			map1["disable_flag"] = fmt.Sprintf("%v", val1["disableFlag"])
+			map1["check_id"] = fmt.Sprintf("%v", val1["checkId"])
+			rcdflist = append(rcdflist, map1)
+		}
+	}
+
+	d.SetId(fmt.Sprintf("%.0f", data["id"]))
+	d.Set("host", data["host"])
+	d.Set("name", data["name"])
+	d.Set("domain_id", params[1])
+	d.Set("source_type", params[0])
+	d.Set("ttl", data["ttl"])
+	d.Set("geo_location", geoset)
+	d.Set("record_option", data["recordOption"])
+	d.Set("noanswer", data["noAnswer"])
+	d.Set("note", data["note"])
+	d.Set("gtd_region", data["gtdRegion"])
+	d.Set("type", data["type"])
+	d.Set("pools", data["pools"])
+	d.Set("contact_ids", data["contactId"])
+	d.Set("record_failover_values", rcdflist)
+	d.Set("record_failover_failover_type", rcdfset["record_failover_failover_type"])
+	d.Set("record_failover_disable_flag", rcdfset["record_failover_disable_flag"])
+	log.Printf("[DEBUG] %s finished import", d.Id())
+	return []*schema.ResourceData{d}, nil
 }
 
 func resourceConstellixCNameRecordCreate(d *schema.ResourceData, m interface{}) error {
@@ -322,7 +405,7 @@ func resourceConstellixCNameRecordRead(d *schema.ResourceData, m interface{}) er
 		}
 	}
 
-	d.Set("id", data["id"])
+	d.SetId(fmt.Sprintf("%.0f", data["id"]))
 	d.Set("host", data["host"])
 	d.Set("name", data["name"])
 	d.Set("ttl", data["ttl"])