+++ constellix/resource_constellix_mx_record.go
@@ -4,6 +4,8 @@ import (
 	"encoding/json"
 	"fmt"
 	"io/ioutil"
+	"log"
+	"strings"
 
 	"github.com/Constellix/constellix-go-client/client"
 	"github.com/Constellix/constellix-go-client/models"
@@ -16,6 +18,10 @@ func resourceConstellixMX() *schema.Resource {
 		Update: resourceConstellixMXUpdate,
 		Read:   resourceConstellixMXRead,
 		Delete: resourceConstellixMXDelete,
+
+		Importer: &schema.ResourceImporter{
+			State: resourceConstellixMXImport,
+		},
 
 		Schema: map[string]*schema.Schema{
 			"domain_id": &schema.Schema{
@@ -88,6 +94,55 @@ func resourceConstellixMX() *schema.Resource {
 			},
 		},
 	}
+}
+
+func resourceConstellixMXImport(d *schema.ResourceData, m interface{}) ([]*schema.ResourceData, error) {
+	log.Printf("[DEBUG] %s: Beginning Import", d.Id())
+	constellixClient := m.(*client.Client)
+	params := strings.Split(d.Id(), ":")
+	resp, err := constellixClient.GetbyId("v1/" + params[0] + "/" + params[1] + "/records/mx/" + params[2])
+	if err != nil {
+		if resp.StatusCode == 404 {
+			d.SetId("")
+			return nil, err
+		}
+		return nil, err
+	}
+	bodybytes, err := ioutil.ReadAll(resp.Body)
+	if err != nil {
+		return nil, err
+	}
+	bodystring := string(bodybytes)
+
+	var data map[string]interface{}
+	json.Unmarshal([]byte(bodystring), &data)
+	d.SetId(fmt.Sprintf("%.0f", data["id"]))
+	d.Set("name", data["name"])
+	d.Set("ttl", data["ttl"])
+	d.Set("noanswer", data["noAnswer"])
+	d.Set("note", data["note"])
+	d.Set("gtd_region", data["gtdRegion"])
+	d.Set("type", data["type"])
+	d.Set("parentid", data["parentId"])
+	d.Set("parent", data["parent"])
+	d.Set("source", data["source"])
+
+	resrr := (data["roundRobin"]).([]interface{})
+	mapListRR := make([]interface{}, 0, 1)
+	for _, val := range resrr {
+		tpMap := make(map[string]interface{})
+		inner := val.(map[string]interface{})
+		tpMap["value"] = fmt.Sprintf("%v", inner["value"])
+		tpMap["level"] = fmt.Sprintf("%v", inner["level"])
+		tpMap["disable_flag"] = fmt.Sprintf("%v", inner["disableFlag"])
+		mapListRR = append(mapListRR, tpMap)
+	}
+
+	d.Set("roundrobin", mapListRR)
+	d.Set("domain_id", params[1])
+	d.Set("source_type", params[0])
+	log.Printf("[DEBUG] %s finished import", d.Id())
+	return []*schema.ResourceData{d}, nil
 }
 
 func resourceConstellixMXCreate(d *schema.ResourceData, m interface{}) error {
@@ -229,7 +284,7 @@ func resourceConstellixMXRead(d *schema.ResourceData, m interface{}) error {
 
 	var data map[string]interface{}
 	json.Unmarshal([]byte(bodystring), &data)
-	d.Set("id", data["id"])
+	d.SetId(fmt.Sprintf("%.0f", data["id"]))
 	d.Set("name", data["name"])
 	d.Set("ttl", data["ttl"])
 	d.Set("noanswer", data["noAnswer"])