+++ constellix/resource_constellix_srv_record.go
@@ -4,7 +4,9 @@ import (
 	"encoding/json"
 	"fmt"
 	"io/ioutil"
+	"log"
 	"strconv"
+	"strings"
 
 	"github.com/Constellix/constellix-go-client/client"
 	"github.com/Constellix/constellix-go-client/models"
@@ -18,6 +20,10 @@ func resourceConstellixSRVRecord() *schema.Resource {
 		Update:        resourceConstellixSRVRecordUpdate,
 		Delete:        resourceConstellixSRVRecordDelete,
 		SchemaVersion: 1,
+
+		Importer: &schema.ResourceImporter{
+			State: resourceConstellixSRVRecordImport,
+		},
 
 		Schema: map[string]*schema.Schema{
 			"domain_id": &schema.Schema{
@@ -105,6 +111,54 @@ func resourceConstellixSRVRecord() *schema.Resource {
 	}
 }
 
+func resourceConstellixSRVRecordImport(d *schema.ResourceData, m interface{}) ([]*schema.ResourceData, error) {
+	log.Printf("[DEBUG] %s: Beginning Import", d.Id())
+	constellixClient := m.(*client.Client)
+	params := strings.Split(d.Id(), ":")
+	resp, err := constellixClient.GetbyId("v1/" + params[0] + "/" + params[1] + "/records/srv/" + params[2])
+	if err != nil {
+		if resp.StatusCode == 404 {
+			d.SetId("")
+			return nil, err
+		}
+		return nil, err
+	}
+	bodyBytes, err := ioutil.ReadAll(resp.Body)
+	if err != nil {
+		return nil, err
+	}
+	bodyString := string(bodyBytes)
+	var data map[string]interface{}
+	err = json.Unmarshal([]byte(bodyString), &data)
+	if err != nil {
+		return nil, err
+	}
+	arecroundrobin := data["roundRobin"].([]interface{})
+	rrlist := make([]interface{}, 0, 1)
+	for _, valrrf := range arecroundrobin {
+		map1 := make(map[string]interface{})
+		val1 := valrrf.(map[string]interface{})
+		map1["value"] = fmt.Sprintf("%v", val1["value"])
+		map1["disable_flag"] = fmt.Sprintf("%v", val1["disableFlag"])
+		map1["port"], _ = strconv.Atoi(fmt.Sprintf("%v", val1["port"]))
+		map1["priority"], _ = strconv.Atoi(fmt.Sprintf("%v", val1["priority"]))
+		map1["weight"], _ = strconv.Atoi(fmt.Sprintf("%v", val1["weight"]))
+		rrlist = append(rrlist, map1)
+	}
+
+	d.SetId(fmt.Sprintf("%.0f", data["id"]))
+	d.Set("name", data["name"])
+	d.Set("ttl", data["ttl"])
+	d.Set("noanswer", data["noAnswer"])
+	d.Set("note", data["note"])
+	d.Set("gtd_region", data["gtdRegion"])
+	d.Set("type", data["type"])
+	d.Set("roundrobin", rrlist)
+	d.Set("domain_id", params[1])
+	d.Set("source_type", params[0])
+	log.Printf("[DEBUG] %s finished import", d.Id())
+	return []*schema.ResourceData{d}, nil
+}
 func resourceConstellixSRVRecordCreate(d *schema.ResourceData, m interface{}) error {
 	constellixConnect := m.(*client.Client)
 	srvAttr := models.SRVAttributes{}
@@ -200,7 +254,7 @@ func resourceConstellixSRVRecordRead(d *schema.ResourceData, m interface{}) erro
 		rrlist = append(rrlist, map1)
 	}
 
-	d.Set("id", data["id"])
+	d.SetId(fmt.Sprintf("%.0f", data["id"]))
 	d.Set("name", data["name"])
 	d.Set("ttl", data["ttl"])
 	d.Set("noanswer", data["noAnswer"])