+++ constellix/resource_constellix_tcp_check.go
@@ -4,6 +4,7 @@ import (
 	"encoding/json"
 	"fmt"
 	"io/ioutil"
+	"log"
 	"strings"
 
 	"github.com/Constellix/constellix-go-client/client"
@@ -18,6 +19,10 @@ func resourceConstellixTCPCheck() *schema.Resource {
 		Update: resourceConstellixTCPCheckUpdate,
 		Read:   resourceConstellixTCPCheckRead,
 		Delete: resourceConstellixTCPCheckDelete,
+
+		Importer: &schema.ResourceImporter{
+			State: resourceConstellixTCPCheckImport,
+		},
 
 		Schema: map[string]*schema.Schema{
 			"name": &schema.Schema{
@@ -93,6 +98,37 @@ func resourceConstellixTCPCheck() *schema.Resource {
 			},
 		},
 	}
+}
+
+func resourceConstellixTCPCheckImport(d *schema.ResourceData, m interface{}) ([]*schema.ResourceData, error) {
+	log.Printf("[DEBUG] %s: Beginning Import", d.Id())
+	constellixClient := m.(*client.Client)
+	dn := d.Id()
+
+	resp, err := constellixClient.GetbyId("https://api.sonar.constellix.com/rest/api/tcp/" + dn)
+	if err != nil {
+		return nil, err
+	}
+	bodybytes, err := ioutil.ReadAll(resp.Body)
+	if err != nil {
+		return nil, err
+	}
+	bodystring := string(bodybytes)
+
+	var data map[string]interface{}
+	json.Unmarshal([]byte(bodystring), &data)
+	d.SetId(fmt.Sprintf("%.0f", data["id"]))
+	d.Set("name", data["name"])
+	d.Set("host", data["host"])
+	d.Set("ip_version", data["ipVersion"])
+	d.Set("port", data["port"])
+	d.Set("check_sites", data["checkSites"])
+	d.Set("interval", data["interval"])
+	d.Set("interval_policy", data["monitorIntervalPolicy"])
+	d.Set("string_to_send", data["stringToSend"])
+	d.Set("string_to_receive", data["stringToReceive"])
+	log.Printf("[DEBUG] %s finished import", d.Id())
+	return []*schema.ResourceData{d}, nil
 }
 
 func resourceConstellixTCPCheckCreate(d *schema.ResourceData, m interface{}) error {
@@ -222,7 +258,7 @@ func resourceConstellixTCPCheckRead(d *schema.ResourceData, m interface{}) error
 
 	var data map[string]interface{}
 	json.Unmarshal([]byte(bodystring), &data)
-	d.Set("id", data["id"])
+	d.SetId(fmt.Sprintf("%.0f", data["id"]))
 	d.Set("name", data["name"])
 	d.Set("host", data["host"])
 	d.Set("ip_version", data["ipVersion"])